+++ packages/utils/rollup.config.js
@@ -1,5 +1,5 @@
 const {
   getUtilRollupConfigurations,
-} = require("@krill/web-framework-tool-rollup-config-factory");
+} = require("@krills/web-framework-tool-rollup-config-factory");
 
 module.exports = getUtilRollupConfigurations();