+++ tools/rollup-config-factory/CHANGELOG.md
@@ -9,7 +9,7 @@
 ### Patch Changes
 
 - Updated dependencies
-  - @krill/web-framework-tool-build-metadata@0.1.0
+  - @krills/web-framework-tool-build-metadata@0.1.0
 
 ## 0.1.0
 
@@ -20,7 +20,7 @@
 ### Patch Changes
 
 - Updated dependencies
-  - @krill/web-framework-tool-build-metadata@0.0.2
+  - @krills/web-framework-tool-build-metadata@0.0.2
 
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.